Fix skillMap insert operation compile error
SkillFunction named function pointer is moved to inside of the CloneCommonSkillFunction. Then, in the Utilities.cpp, SkillFunction is typedef'ed. Then all the problems are solved. But I chose to initialize "skillFunction" in the header. Because it works as a look up table. I noted this statement in the Utilities.cpp also.