Define get_Index and get_FromIndex functions
The enum class is used for weapons, skills, soldier types and so on for this project.
For, mathematic operations such as damage or printing operations such as printing a skill name to the screen, get_  and get_Str methods are used.
* get_ returns the enum object
* get_Str returns its string version.
But for saving and loading operations, writing all the string for identification is unnecessary, Writing the index for that particular enum is very sufficient. For this reason, the enum class object must be staticly casted into int typed number.
Because, there are too much enum classes, lots of casting operations took place in other classes, that uses classes comprimise an enum class. Hence two new get functions created.
* get_Index
* get_FromIndex
get_Index functions returns, the index of that enum. Casting operation takes place inside of this function.
get_FromIndex function on the other hand, return the enum given from index. Casting operation is the reverse of that used in get_Index function.
Both of the functions are defined as static.